moved test, added asserts
